Adding support to skip tests that will always fail.
Many of the online tests always fail because the services have
not been implemented, which causes a lot of noise and makes
running the tests and looking for meaningful failures a hassle.

I've added support so that a container-specific test suite can
opt out of a unit test by declaring which osapi methods it 
explicitly does not support.  Online unit tests must then
assert the names of services they use.

Eventually this kind of functionality would be best served as
being part of the provider classes and not directly in the tests
themselves, but that will call for a bit of refactoring before
it will be possible.