Made use of the "body hack" method of OAuth signing configurable.
Some OAuth libraries (incorrectly) looked at a non
x-www-form-urlencoded POST body as a single parameter that needed
to be included in the request when signing.  To maintain
compatibility with containers running such clients, this
functionality was introduced into the PHP client library when
signing a request.

Most containers should have updated OAuth libraries now, and
most likely won't need this workaround.  It also becomes
a liability when working with non application/json requests
(such as the recent addition of MediaItem support).

This patch disables body hack signing by default.  For
containers which still require this, the following code can
be used in the provider's preRequestProcess handler:

  if (method_exists($signer, 'setUseBodyHack')) {
    $signer->setUseBodyHack(true);
  }

Note that there is a difference between "setUseBodyHack"
and "setUseBodyHash" although both can be used together.

All offline tests pass.